docs(storybook): refactor PayPalButtons and Braintree
