+++ package.json
@@ -70,6 +70,7 @@
         "@types/jest": "^27.0.2",
         "@types/react": "^17.0.25",
         "@types/react-dom": "^17.0.9",
+        "@types/string-template": "^1.0.2",
         "@typescript-eslint/eslint-plugin": "^4.32.0",
         "@typescript-eslint/parser": "^4.32.0",
         "babel-jest": "^27.2.4",
@@ -97,6 +98,7 @@
         "rollup-plugin-terser": "^7.0.2",
         "shelljs": "^0.8.4",
         "standard-version": "^9.3.1",
+        "string-template": "^1.0.0",
         "typescript": "^4.4.3"
     },
     "peerDependencies": {
