+++ src/stories/braintree/BraintreePayPalButtons.stories.tsx
@@ -28,7 +28,7 @@ import {
     APPROVE,
 } from "../constants";
 import { InEligibleError, defaultProps } from "../commons";
-import overrideStories from "./utils";
+import overrideStories from "./code";
 
 type StoryProps = {
     style: PayPalButtonsComponentOptions["style"];
