+++ src/stories/payPalButtons/PayPalButtons.stories.tsx
@@ -21,7 +21,7 @@ import {
     APPROVE,
 } from "../constants";
 import { InEligibleError, defaultProps } from "../commons";
-import overrideStories from "./utils";
+import overrideStories from "./code";
 
 type StoryProps = {
     style: PayPalButtonsComponentOptions["style"];
